requirements.txt
ecdsa==0.18.0
PyYAML==6.0
six==1.16.0
